requirements.txt
Flask
MarkupSafe
itsdangerous
Werkzeug
click
python-dotenv
alchemical
SQLAlchemy
flask-wtf
wtforms
flask-minify